Move serializer and emote option to configuration package
