+++ connector/src/main/java/org/geysermc/connector/configuration/GeyserJacksonConfiguration.java
@@ -32,8 +32,6 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import lombok.Getter;
 import lombok.Setter;
 import org.geysermc.connector.GeyserConnector;
-import org.geysermc.connector.common.EmoteOffhandDeserializer;
-import org.geysermc.connector.common.EmoteOffhandWorkaroundOption;
 import org.geysermc.connector.common.serializer.AsteriskSerializer;
 import org.geysermc.connector.network.CIDRMatcher;
 
@@ -103,7 +101,7 @@ public abstract class GeyserJacksonConfiguration implements GeyserConfiguration 
     @JsonProperty("show-coordinates")
     private boolean showCoordinates = true;
 
-    @JsonDeserialize(using = EmoteOffhandDeserializer.class)
+    @JsonDeserialize(using = EmoteOffhandWorkaroundOption.EmoteOffhandDeserializer.class)
     @JsonProperty("emote-offhand-workaround")
     private EmoteOffhandWorkaroundOption emoteOffhandWorkaround = EmoteOffhandWorkaroundOption.DISABLED;
