+++ connector/src/main/resources/config.yml
@@ -127,7 +127,7 @@ show-coordinates: true
 # If set, when a Bedrock player performs any emote, it will swap the offhand and mainhand items, just like the Java Edition keybind
 # There are three options this can be set to:
 # disabled - the default/fallback, which doesn't apply this workaround
-# no-emotes - emotes will not be sent to other Bedrock clients. This effectively disables all emotes from being seen.
+# no-emotes - emotes will NOT be sent to other Bedrock clients and offhand will be swapped. This effectively disables all emotes from being seen.
 # emotes-and-offhand - emotes will be sent to Bedrock clients and offhand will be swapped
 emote-offhand-workaround: "disabled"
